+++ packages/shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./api-client.js";
 export * from "./constants.js";
+export * from "./email.js";
 export * from "./env.js";
 export * from "./logger.js";
 export * from "./storage.js";